Keep shared mock in log_processor
